Restoring ability to rebrand via LESS change and fixing styling of system-status
Steve had used an image to add branding instead of a background-image.
This is problematic as we rebrand for OCP using LESS changes.

Note:  we used to have TWO nav-system-status extensions
(nav-system-status and nav-system-status-mobile).  With the new design,
we no longer have a need for nav-system-status-mobile, so I’ve removed
it).  This should not pose a problem as definitions for
nav-system-status-mobile in an extensions.js file with will be ignored.
In the openshiftOnlineExtentions module, nav-system-status-mobile was a
clone of nav-system-status, so no unique functionality is lost at least
for Online.  It’s theoretically possible someone was using the two
distinct nav-system-status-es for different purposes, but it seems
highly unlikely.